Refactor: 그룹 가입 관련 기존 코드 이동
Ref: #72
Group -> GroupEntry
