+++ src/main/java/com/studypals/domain/groupManage/api/GroupController.java
@@ -13,8 +13,6 @@ import lombok.RequiredArgsConstructor;
 
 import com.studypals.domain.groupManage.dto.CreateGroupReq;
 import com.studypals.domain.groupManage.dto.GetGroupTagRes;
-import com.studypals.domain.groupManage.dto.GroupEntryCodeRes;
-import com.studypals.domain.groupManage.dto.GroupSummaryRes;
 import com.studypals.domain.groupManage.service.GroupService;
 import com.studypals.global.responses.CommonResponse;
 import com.studypals.global.responses.Response;
@@ -26,8 +24,6 @@ import com.studypals.global.responses.ResponseCode;
  * <pre>
  *     - GET /groups/tags : 그룹 태그 조회
  *     - POST /groups : 그룹 생성({@link CreateGroupReq})
- *     - POST /groups/{groupId}/entry-code : 그룹 초대 코드 생성
- *     - GET /groups/summary : 그룹 대표 정보 조회
  * </pre>
  *
  * @author s0o0bn
@@ -52,21 +48,5 @@ public class GroupController {
         Long groupId = groupService.createGroup(userId, request);
 
         return ResponseEntity.created(URI.create("/groups/" + groupId)).build();
-    }
-
-    @PostMapping("/{groupId}/entry-code")
-    public ResponseEntity<Response<GroupEntryCodeRes>> generateEntryCode(
-            @AuthenticationPrincipal Long userId, @PathVariable Long groupId) {
-        GroupEntryCodeRes codeResponse = groupService.generateEntryCode(userId, groupId);
-        Response<GroupEntryCodeRes> response = CommonResponse.success(ResponseCode.GROUP_ENTRY_CODE, codeResponse);
-
-        return ResponseEntity.created(URI.create("/groups/" + groupId + "/entry-code/" + codeResponse.code()))
-                .body(response);
-    }
-
-    @GetMapping("/summary")
-    public ResponseEntity<Response<GroupSummaryRes>> getGroupSummary(@RequestParam String entryCode) {
-        GroupSummaryRes response = groupService.getGroupSummary(entryCode);
-        return ResponseEntity.ok(CommonResponse.success(ResponseCode.GROUP_SUMMARY, response));
     }
 }
