+++ src/main/java/com/studypals/domain/groupManage/service/GroupEntryServiceImpl.java
@@ -1,16 +1,18 @@
 package com.studypals.domain.groupManage.service;
 
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 
+import com.studypals.domain.groupManage.dto.GroupEntryCodeRes;
 import com.studypals.domain.groupManage.dto.GroupEntryReq;
+import com.studypals.domain.groupManage.dto.GroupMemberProfileDto;
+import com.studypals.domain.groupManage.dto.GroupSummaryRes;
 import com.studypals.domain.groupManage.entity.Group;
-import com.studypals.domain.groupManage.worker.GroupEntryCodeManager;
-import com.studypals.domain.groupManage.worker.GroupEntryRequestWorker;
-import com.studypals.domain.groupManage.worker.GroupMemberWorker;
-import com.studypals.domain.groupManage.worker.GroupReader;
+import com.studypals.domain.groupManage.worker.*;
 import com.studypals.domain.memberManage.entity.Member;
 import com.studypals.domain.memberManage.worker.MemberReader;
 import com.studypals.global.exceptions.errorCode.GroupErrorCode;
@@ -34,11 +36,37 @@ import com.studypals.global.exceptions.exception.GroupException;
 @Service
 @RequiredArgsConstructor
 public class GroupEntryServiceImpl implements GroupEntryService {
+    private static final int GROUP_SUMMARY_MEMBER_COUNT = 5;
+
     private final MemberReader memberReader;
     private final GroupReader groupReader;
     private final GroupMemberWorker groupMemberWorker;
+    private final GroupMemberReader groupMemberReader;
+
+    private final GroupAuthorityValidator authorityValidator;
     private final GroupEntryCodeManager entryCodeManager;
     private final GroupEntryRequestWorker entryRequestWorker;
+
+    @Override
+    @Transactional(readOnly = true)
+    public GroupEntryCodeRes generateEntryCode(Long userId, Long groupId) {
+        Group group = groupReader.getById(groupId);
+        authorityValidator.validate(userId, groupId);
+        String entryCode = entryCodeManager.generate(group.getId());
+
+        return new GroupEntryCodeRes(group.getId(), entryCode);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public GroupSummaryRes getGroupSummary(String entryCode) {
+        Long groupId = entryCodeManager.getGroupId(entryCode);
+        Group group = groupReader.getById(groupId);
+        List<GroupMemberProfileDto> profiles =
+                groupMemberReader.getTopNMemberProfiles(groupId, GROUP_SUMMARY_MEMBER_COUNT);
+
+        return GroupSummaryRes.of(group, profiles);
+    }
 
     @Override
     @Transactional
