+++ src/main/java/com/studypals/domain/groupManage/service/GroupService.java
@@ -4,8 +4,6 @@ import java.util.List;
 
 import com.studypals.domain.groupManage.dto.CreateGroupReq;
 import com.studypals.domain.groupManage.dto.GetGroupTagRes;
-import com.studypals.domain.groupManage.dto.GroupEntryCodeRes;
-import com.studypals.domain.groupManage.dto.GroupSummaryRes;
 
 /**
  * GroupService 의 인터페이스입니다. 메서드를 정의합니다.
@@ -37,24 +35,4 @@ public interface GroupService {
      * @throws com.studypals.global.exceptions.exception.GroupException
      */
     Long createGroup(Long userId, CreateGroupReq dto);
-
-    /**
-     * 그룹 초대 코드를 생성한다.
-     * 코드 생성은 그룹장만 가능하다.
-     *
-     * @param userId 초대 코드를 생성한 사용자 ID
-     * @param groupId 코드를 생성할 그룹 ID
-     * @return 그룹 ID와 해당 그룹의 초대 코드
-     */
-    GroupEntryCodeRes generateEntryCode(Long userId, Long groupId);
-
-    /**
-     * 초대 코드로 그룹 대표 정보를 조회합니다.
-     * 그룹장 포함 일부 멤버들의 프로필 이미지와 함께 조회합니다.
-     *
-     * @param entryCode 그룹 초대 코드
-     * @return 그룹 대표 정보
-     * @throws com.studypals.global.exceptions.exception.GroupException
-     */
-    GroupSummaryRes getGroupSummary(String entryCode);
 }
